Move to Shared -SqlDataRecord
this is part of #1261 issue code merge classes. Moved the SqlDataRecord common code to .Common.cs class and insternally invoked private methods from corresponsding FrameworkSpecific methods.
